Update version number to reflect NPE fix.
Reorganize test into separate class and smaller files.

Address localization in Test.
