+++ Project/src/test/java/TestFeatures.java
@@ -21,13 +21,7 @@
  * THE SOFTWARE.
  */
 
-import com.github.lgooddatepicker.components.CalendarPanel;
-import com.github.lgooddatepicker.components.DatePicker;
-import com.github.lgooddatepicker.components.DatePickerSettings;
-import com.github.lgooddatepicker.components.TimePicker;
-import com.github.lgooddatepicker.components.TimePickerSettings;
-import com.github.lgooddatepicker.optionalusertools.TimeChangeListener;
-import com.github.lgooddatepicker.zinternaltools.TimeChangeEvent;
+import static org.junit.Assert.assertTrue;
 
 import java.awt.Color;
 import java.awt.event.MouseEvent;
@@ -40,11 +34,17 @@ import java.time.Month;
 import java.time.YearMonth;
 import java.time.ZoneId;
 import java.time.ZoneOffset;
-import java.time.temporal.ChronoUnit;
 import java.util.Locale;
+
 import javax.swing.JLabel;
+
 import org.junit.Test;
-import static org.junit.Assert.*;
+
+import com.github.lgooddatepicker.components.CalendarPanel;
+import com.github.lgooddatepicker.components.DatePicker;
+import com.github.lgooddatepicker.components.DatePickerSettings;
+import com.github.lgooddatepicker.components.TimePicker;
+import com.github.lgooddatepicker.components.TimePickerSettings;
 
 // add tests for your new features here
 public class TestFeatures
@@ -184,111 +184,7 @@ public class TestFeatures
         assertTrue(labelname+" has wrong text color: "+labeltoverify.getForeground().toString(), labeltoverify.getForeground().equals(defaultText));
     }
 
-    @Test( expected = Test.None.class /* no exception expected */ )
-    public void verifyTimePicker()
-    {
-        TimePicker picker = new TimePicker();
-        picker.setTime(LocalTime.MIN);
-        assertEquals("minium local time could not be used", LocalTime.MIN, picker.getTime());
-        picker.setTime(LocalTime.NOON);
-        assertEquals("noon local time could not be used", LocalTime.NOON, picker.getTime());
-        picker.setTime(LocalTime.MAX);
-        assertEquals("maximum local time could not be used", LocalTime.MAX.truncatedTo(ChronoUnit.MINUTES), picker.getTime());
-        picker.setTime(null);
-        assertNull("null time could not be used", picker.getTime());
-        picker.setEnableArrowKeys(true);
-        assertTrue("Arrow keys not enabled", picker.getEnableArrowKeys());
-        picker.setEnableArrowKeys(false);
-        assertFalse("Arrow keys not disabled", picker.getEnableArrowKeys());
-        
-        assertNotNull("Picker settings were null", picker.getSettings());       
-        
-        picker.setEnabled(false);
-        assertFalse("Picker was not disabled", picker.isEnabled());
-        assertFalse("Menu component was not disabled", picker.getComponentToggleTimeMenuButton().isEnabled());
-        assertFalse("TextField component was not disabled", picker.getComponentTimeTextField().isEnabled());
-        
-        picker.setEnabled(true);
-        assertTrue("Picker was not disabled", picker.isEnabled());
-        assertTrue("Menu component was not enabled", picker.getComponentToggleTimeMenuButton().isEnabled());
-        assertTrue("TextField component was not enabled", picker.getComponentTimeTextField().isEnabled());
-        
-        picker.setTime(LocalTime.NOON);
-        assertEquals("noon local time could not be used", LocalTime.NOON, picker.getTime());
-        picker.clear();
-        assertNull("Clear did not make the time null", picker.getTime());
-        
-        // valid text
-        picker.setText("12:22");
-        assertTrue("Expected time to be valid", picker.isTextValid("12:22"));
-        assertTrue("Expected field to be valid", picker.isTextFieldValid());
-        assertEquals("Did not retain user text", "12:22", picker.getText());
-        assertEquals("Entered time not translated to local time", LocalTime.of(12, 22), picker.getTime());
-        assertEquals("Expected time string for valid time", "12:22", picker.getTimeStringOrSuppliedString("supply"));
-        
-        // invalid text
-        picker.setText("44:17");
-        assertFalse("Expected time to be invalid", picker.isTextValid("44:17"));
-        assertFalse("Expected timefield  to be invalid", picker.isTextFieldValid());
-        assertEquals("Did not retain user text", "44:17", picker.getText());
-        // because time is invalid the old local time should still be present
-        assertEquals("Invalid time was translated to local time", LocalTime.of(12, 22), picker.getTime());
-        
-        // null time
-        picker.setTime(null);
-        assertEquals("Expected empty string for null time", "", picker.getTimeStringOrEmptyString());
-        assertEquals("Expected supplied string for null time", "supply", picker.getTimeStringOrSuppliedString("supply"));
-        
-        // null text
-        assertFalse("null text was considered valid", picker.isTextValid(null));
-        picker.setText(null);
-        assertEquals("null text did not become blank text", "", picker.getText());
-        
-        // empty text
-        assertTrue("spaces only text was considered valid", picker.isTextValid("  "));
-        picker.setText("  ");
-        assertEquals("spaces text was not returned", "  ", picker.getText());
-        
-        // toString
-        picker.setTime(LocalTime.of(8, 32));
-        assertEquals("toString should match toTime", picker.getTimeStringOrEmptyString(), picker.toString());
-        picker.setTime(LocalTime.of(8, 32, 12));
-        assertEquals("toString should match toTime", picker.getTimeStringOrEmptyString(), picker.toString());
-        
-        assertTrue("Expect noon to be an allowed time", picker.isTimeAllowed(LocalTime.NOON));
-        
-    }
-
-    @Test( expected = Test.None.class /* no exception expected */ )
-    public void verifyTimeChangeListeners()
-    {
-        TimePicker picker = new TimePicker();
-        TestableTimeChangeListener listener = new TestableTimeChangeListener();
-        picker.addTimeChangeListener(listener);
-        assertNull("listener event not null at start", listener.getLastEvent());
-        picker.setTime(LocalTime.MIN);
-        assertEquals("Listener did not receive new time", LocalTime.MIN, listener.getLastEvent().getNewTime());
-        assertNull("Listener did not remember old time", listener.getLastEvent().getOldTime());
-        assertEquals("Event did not originate from time picker", picker, listener.getLastEvent().getSource());
-
-        TimeChangeEvent lastEvent = listener.getLastEvent();
-        picker.setTime(LocalTime.MIN);
-        assertTrue("Event updated when time did not change", lastEvent == listener.getLastEvent());
-
-        picker.setTime(LocalTime.NOON);
-        assertEquals("Listener did not remember old time", LocalTime.MIN, listener.getLastEvent().getOldTime());
-        assertEquals("Listener did not receive new time", LocalTime.NOON, listener.getLastEvent().getNewTime());
-
-        picker.setTime(null);
-        assertNull("Listener did not receive null time", listener.getLastEvent().getNewTime());
-        
-        assertTrue("Listener was not in the list of listeners", picker.getTimeChangeListeners().contains(listener));
-        
-        picker.removeTimeChangeListener(listener);
-        picker.setTime(LocalTime.NOON);
-        assertNull("Listener received an update after being uninstalled", listener.getLastEvent().getNewTime());
-        
-    }
+   
 
     // helper functions
     Clock getClockFixedToInstant(int year, Month month, int day, int hours, int minutes)
@@ -311,18 +207,5 @@ public class TestFeatures
         return private_method;
     }
 
-    //helper class
-    private class TestableTimeChangeListener implements TimeChangeListener
-    {
-        TimeChangeEvent lastEvent;
-
-        @Override
-        public void timeChanged(TimeChangeEvent event) {
-            lastEvent = event;
-        }
-
-        TimeChangeEvent getLastEvent() {
-            return lastEvent;
-        }
-    }
+   
 }
